+++ lib/adhearsion/call.rb
@@ -1,11 +1,30 @@
 require 'thread'
 
+module Celluloid
+  module ClassMethods
+    def ===(other)
+      other.kind_of? self
+    end
+  end
+
+  class ActorProxy
+    def is_a?(klass)
+      Actor.call @mailbox, :is_a?, klass
+    end
+
+    def kind_of?(klass)
+      Actor.call @mailbox, :kind_of?, klass
+    end
+  end
+end
+
 module Adhearsion
   ##
   # Encapsulates call-related data and behavior.
   #
   class Call
 
+    include Celluloid
     include HasGuardedHandlers
 
     attr_accessor :offer, :client, :end_reason, :commands, :variables
@@ -16,39 +35,35 @@ module Adhearsion
     def initialize(offer = nil)
       register_initial_handlers
 
-      @tag_mutex        = Mutex.new
-      @tags             = []
-      @end_reason_mutex = Mutex.new
-      @commands         = CommandRegistry.new
-      @variables        = {}
+      @tags       = []
+      @commands   = CommandRegistry.new
+      @variables  = {}
 
       self << offer if offer
     end
 
     def id
-      offer.call_id
+      offer.call_id if offer
     end
 
     def tags
-      @tag_mutex.synchronize { @tags.clone }
+      @tags.clone
     end
 
     # This may still be a symbol, but no longer requires the tag to be a symbol although beware
     # that using a symbol would create a memory leak if used improperly
     # @param [String, Symbol] label String or Symbol with which to tag this call
     def tag(label)
-      raise ArgumentError, "Tag must be a String or Symbol" unless [String, Symbol].include?(label.class)
-      @tag_mutex.synchronize { @tags << label }
+      abort ArgumentError.new "Tag must be a String or Symbol" unless [String, Symbol].include?(label.class)
+      @tags << label
     end
 
     def remove_tag(symbol)
-      @tag_mutex.synchronize do
-        @tags.reject! { |tag| tag == symbol }
-      end
+      @tags.reject! { |tag| tag == symbol }
     end
 
     def tagged_with?(symbol)
-      @tag_mutex.synchronize { @tags.include? symbol }
+      @tags.include? symbol
     end
 
     def register_event_handler(*guards, &block)
@@ -56,6 +71,7 @@ module Adhearsion
     end
 
     def deliver_message(message)
+      logger.debug "Receiving message: #{message.inspect}"
       trigger_handler :event, message
     end
 
@@ -74,9 +90,10 @@ module Adhearsion
       end
 
       on_end do |event|
-        hangup
-        @end_reason_mutex.synchronize { @end_reason = event.reason }
+        clear_from_active_calls
+        @end_reason = event.reason
         commands.terminate
+        after(5) { current_actor.terminate! }
       end
     end
 
@@ -88,11 +105,11 @@ module Adhearsion
     end
 
     def active?
-      @end_reason_mutex.synchronize { !end_reason }
+      !end_reason
     end
 
     def accept(headers = nil)
-      write_and_await_response Punchblock::Command::Accept.new(:headers => headers)
+      @accept_command ||= write_and_await_response Punchblock::Command::Accept.new(:headers => headers)
     end
 
     def answer(headers = nil)
@@ -103,14 +120,14 @@ module Adhearsion
       write_and_await_response Punchblock::Command::Reject.new(:reason => reason, :headers => headers)
     end
 
-    def hangup!(headers = nil)
+    def hangup(headers = nil)
       return false unless active?
-      @end_reason_mutex.synchronize { @end_reason = true }
+      @end_reason = true
       write_and_await_response Punchblock::Command::Hangup.new(:headers => headers)
     end
 
-    def hangup
-      Adhearsion.active_calls.remove_inactive_call self
+    def clear_from_active_calls
+      Adhearsion.active_calls.remove_inactive_call current_actor
     end
 
     ##
@@ -128,7 +145,7 @@ module Adhearsion
       when String
         options[:other_call_id] = target
       when Hash
-        raise ArgumentError, "You cannot specify both a call ID and mixer name" if target.has_key?(:call_id) && target.has_key?(:mixer_name)
+        abort ArgumentError.new "You cannot specify both a call ID and mixer name" if target.has_key?(:call_id) && target.has_key?(:mixer_name)
         target.tap do |t|
           t[:other_call_id] = t[:call_id]
           t.delete :call_id
@@ -136,7 +153,7 @@ module Adhearsion
 
         options.merge! target
       else
-        raise ArgumentError, "Don't know how to join to #{target.inspect}"
+        abort ArgumentError.new "Don't know how to join to #{target.inspect}"
       end
       command = Punchblock::Command::Join.new options
       write_and_await_response command
@@ -150,29 +167,35 @@ module Adhearsion
       write_and_await_response ::Punchblock::Command::Unmute.new
     end
 
-    def with_command_lock
-      @command_monitor ||= Monitor.new
-      @command_monitor.synchronize { yield }
-    end
-
     def write_and_await_response(command, timeout = 60)
-      # TODO: Put this back once we figure out why it's causing CI to fail
-      # logger.trace "Executing command #{command.inspect}"
       commands << command
       write_command command
-      response = command.response timeout
-      raise response if response.is_a? Exception
+      begin
+        response = command.response timeout
+      rescue Timeout::Error => e
+        abort e
+      end
+      abort response if response.is_a? Exception
       command
     end
 
     def write_command(command)
-      raise Hangup unless active? || command.is_a?(Punchblock::Command::Hangup)
+      abort Hangup.new unless active? || command.is_a?(Punchblock::Command::Hangup)
       variables.merge! command.headers_hash if command.respond_to? :headers_hash
+      logger.trace "Executing command #{command.inspect}"
       client.execute_command command, :call_id => id
     end
 
     def logger_id
       "#{self.class}: #{id}"
+    end
+
+    def logger
+      super
+    end
+
+    def to_ary
+      [current_actor]
     end
 
     def execute_controller(controller, latch = nil)
@@ -181,7 +204,7 @@ module Adhearsion
           begin
             CallController.exec controller
           ensure
-            hangup!
+            hangup
           end
           latch.countdown! if latch
         end
@@ -194,23 +217,6 @@ module Adhearsion
         each { |command| command.response = hangup if command.requested? }
       end
     end
-
-    class Registry
-      @registry = Hash.new
-      @mutex = Mutex.new
-
-      def self.[](k)
-        @mutex.synchronize do
-          @registry[k]
-        end
-      end
-
-      def self.[]=(k, value)
-        @mutex.synchronize do
-          @registry[k] = value
-        end
-      end
-    end#Registry
 
   end#Call
 end#Adhearsion