+++ lib/adhearsion/call_controller.rb
@@ -34,16 +34,13 @@ module Adhearsion
     end
 
     class << self
-      def exec(controller, fresh_call = true)
-        return unless controller
-
+      def exec(controller)
         new_controller = catch :pass_controller do
-          controller.skip_accept! unless fresh_call
           controller.execute!
           nil
         end
 
-        exec new_controller, false
+        exec new_controller if new_controller
       end
 
       ##
@@ -57,7 +54,7 @@ module Adhearsion
 
     delegate :[], :[]=, :to => :@metadata
     delegate :variables, :logger, :to => :call
-    delegate :write_and_await_response, :accept, :answer, :reject, :mute, :unmute, :join, :to => :call
+    delegate :write_and_await_response, :answer, :reject, :mute, :unmute, :join, :to => :call
 
     def initialize(call, metadata = nil, &block)
       @call, @metadata, @block = call, metadata || {}, block
@@ -65,7 +62,6 @@ module Adhearsion
 
     def execute!(*options)
       execute_callbacks :before_call
-      accept if auto_accept?
       run
     rescue Hangup
       logger.info "Call was hung up"
@@ -96,24 +92,12 @@ module Adhearsion
       end
     end
 
-    def skip_accept!
-      @skip_accept = true
-    end
-
-    def skip_accept?
-      @skip_accept || false
-    end
-
-    def auto_accept?
-      Adhearsion.config.platform.automatically_accept_incoming_calls && !skip_accept?
-    end
-
     def after_call
       @after_call ||= execute_callbacks :after_call
     end
 
     def hangup(headers = nil)
-      hangup_response = call.hangup! headers
+      hangup_response = call.hangup headers
       after_call unless hangup_response == false
     end