+++ lib/adhearsion/console.rb
@@ -2,8 +2,9 @@ require 'pry'
 
 module Adhearsion
   class Console
-    include Adhearsion
     include Singleton
+
+    delegate :silence!, :unsilence!, :to => Adhearsion::Logging
 
     class << self
       ##
@@ -15,6 +16,12 @@ module Adhearsion
       def method_missing(method, *args, &block)
         instance.send method, *args, &block
       end
+    end
+
+    attr_accessor :input
+
+    def initialize
+      @input = $stdin
     end
 
     ##
@@ -48,21 +55,47 @@ module Adhearsion
       logger.info "Shutting down"
     end
 
+    def log_level(level = nil)
+      if level
+        Adhearsion::Logging.level = level
+      else
+        ::Logging::LEVELS.invert[Adhearsion::Logging.level].to_sym
+      end
+    end
+
+    def shutdown!
+      Process.shutdown!
+    end
+
     def calls
       Adhearsion.active_calls
     end
 
-    def use(call)
-      unless call.is_a? Adhearsion::Call
-        raise ArgumentError unless Adhearsion.active_calls[call]
-        call = Adhearsion.active_calls[call]
-      end
-      Pry.prompt = [ proc { "AHN<#{call.channel}> " },
-                     proc { "AHN<#{call.channel}? " }  ]
-
-      # Pause execution of the thread currently controlling the call
-      call.with_command_lock do
-        CallWrapper.new(call).pry
+    def take(call = nil)
+      case call
+      when Call
+        interact_with_call call
+      when String
+        if call = calls[call]
+          interact_with_call call
+        else
+          logger.error "An active call with that ID does not exist"
+        end
+      when nil
+        if calls.size == 1
+          interact_with_call calls.values.first
+        else
+          puts "Please choose a call:"
+          current_calls = calls.values
+          current_calls.each_with_index do |call, index|
+            puts "#{index}. (#{call.is_a?(OutboundCall) ? 'o' : 'i' }) #{call.id} from #{call.from} to #{call.to}"
+          end
+          index = input.gets.chomp.to_i
+          call = current_calls[index]
+          interact_with_call call
+        end
+      else
+        raise ArgumentError
       end
     end
 
@@ -76,12 +109,18 @@ module Adhearsion
       end
     end
 
-    class CallWrapper
-      attr_accessor :call
+    private
 
-      def initialize(call)
-        @call = call
-        extend Adhearsion::Commands.for('asterisk')
+    def interact_with_call(call)
+      Pry.prompt = [ proc { "AHN<#{call.id}> " },
+                     proc { "AHN<#{call.id}? " }  ]
+
+      CallController.exec InteractiveController.new(call)
+    end
+
+    class InteractiveController < CallController
+      def run
+        pry
       end
     end
   end