+++ spec/adhearsion/call_controller_spec.rb
@@ -29,41 +29,7 @@ module Adhearsion
     describe "execution on a call" do
       before do
         flexmock subject, :execute_component_and_await_completion => nil
-        flexmock call, :write_and_await_response => nil
-      end
-
-      context "when auto-accept is enabled" do
-        before do
-          Adhearsion.config.platform.automatically_accept_incoming_calls = true
-        end
-
-        it "should accept the call" do
-          subject.should_receive(:accept).once.ordered
-          subject.should_receive(:run).once.ordered
-          subject.execute!
-        end
-
-        context "and accept is skipped" do
-          before { subject.skip_accept! }
-
-          it "should not accept the call" do
-            subject.should_receive(:accept).never
-            subject.should_receive(:run).once
-            subject.execute!
-          end
-        end
-      end
-
-      context "when auto-accept is disabled" do
-        before do
-          Adhearsion.config.platform.automatically_accept_incoming_calls = false
-        end
-
-        it "should not accept the call" do
-          subject.should_receive(:accept).never
-          subject.should_receive(:run).once
-          subject.execute!
-        end
+        flexmock call.wrapped_object, :write_and_await_response => nil
       end
 
       it "catches Hangup exceptions and logs the hangup" do
@@ -137,9 +103,8 @@ module Adhearsion
 
       before do
         flexmock subject, :execute_component_and_await_completion => nil
-        flexmock call, :write_and_await_response => nil
+        flexmock call.wrapped_object, :write_and_await_response => nil
         flexmock(Events).should_receive(:trigger).with(:exception, Exception).never
-        Adhearsion.config.platform.automatically_accept_incoming_calls = true
       end
 
       it "should invoke another controller before returning to the current controller" do
@@ -152,12 +117,6 @@ module Adhearsion
 
       it "should invoke the new controller with metadata" do
         flexmock(SecondController).new_instances.should_receive(:md_check).once.with :foo => 'bar'
-        subject.execute!
-      end
-
-      it "should not attempt to accept the call again" do
-        call.should_receive(:accept).once
-
         subject.execute!
       end
 
@@ -195,11 +154,10 @@ module Adhearsion
       subject { PassController.new call }
 
       before do
-        flexmock(call).should_receive(:write_and_await_response).and_return nil
+        flexmock(call.wrapped_object).should_receive(:write_and_await_response).and_return nil
         flexmock subject, :execute_component_and_await_completion => nil
         flexmock(SecondController).new_instances.should_receive(:md_check).once.with :foo => 'bar'
         flexmock(Events).should_receive(:trigger).with(:exception, Exception).never
-        Adhearsion.config.platform.automatically_accept_incoming_calls = true
       end
 
       let(:latch) { CountDownLatch.new 1 }
@@ -208,16 +166,10 @@ module Adhearsion
         subject.should_receive(:before).once.ordered
         call.should_receive(:answer).once.ordered
         subject.should_receive(:after).never.ordered
-        call.should_receive(:hangup!).once.ordered
+        call.wrapped_object.should_receive(:hangup).once.ordered
 
         call.execute_controller subject, latch
         latch.wait(1).should be_true
-      end
-
-      it "should not attempt to accept the call again" do
-        call.should_receive(:accept).once
-
-        CallController.exec subject
       end
 
       it "should execute after_call callbacks before passing control" do
@@ -296,13 +248,6 @@ module Adhearsion
       end
     end
 
-    describe '#accept' do
-      it "should delegate to the call" do
-        flexmock(call).should_receive(:accept).once.with(:foo)
-        subject.accept :foo
-      end
-    end
-
     describe '#answer' do
       it "should delegate to the call" do
         flexmock(call).should_receive(:answer).once.with(:foo)
@@ -319,7 +264,7 @@ module Adhearsion
 
     describe '#hangup' do
       it "should delegate to the call" do
-        flexmock(call).should_receive(:hangup!).once.with(:foo)
+        flexmock(call).should_receive(:hangup).once.with(:foo)
         subject.hangup :foo
       end
     end
@@ -378,13 +323,11 @@ describe ExampleCallController do
 
   before do
     flexmock subject, :execute_component_and_await_completion => nil
-    flexmock call, :write_and_await_response => nil
-    Adhearsion.config.platform.automatically_accept_incoming_calls = true
+    flexmock call.wrapped_object, :write_and_await_response => nil
   end
 
-  it "should execute the before_call callbacks before accepting the call" do
+  it "should execute the before_call callbacks before processing the call" do
     subject.should_receive(:setup_models).twice.ordered
-    subject.should_receive(:accept).once.ordered
     subject.should_receive(:join_to_conference).once.ordered
     subject.execute!
   end