+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-# develop (2.0.0.head)
+# 2.0.0.alpha3 - 2012-02-21
+  * Feature: Add `ahn generate` command to allow invocation of generators
+  * Feature: Add simple generator for call controllers
+  * Feature: Add simple generator for plugins
+  * Feature: Allow plugins to register their generator classes
+  * Feature: Add log level helper methods to Console
+  * Feature: Console's shutdown/exit method initiates the shutdown routine
+  * Bugfix: Remove config option for auto-accept - hard-coded to true
+  * Bugfix: AHN_ENV and RAILS_ENV now do not interfere with each other when both are set, and ahn will boot in the RAILS_ENV if AHN_ENV is not set
+  * Feature: The console can take control of a call
+  * Bugfix: CallController#dial now blocks until all outbound calls complete
+  * Bugfix: Call commands timing out now raise a timeout exception in the caller, but do not crash the actor
+  * Bugfix: CallController#dial now unblocks immediately if the original call ends
+  * Bugfix: CallController#dial now unblocks when the connected outbound call unjoins, rather than ending, incase post-processing on the outbound call is required
+  * Bugfix: CallController#dial now hangs up outbound legs when it unblocks
+  * Feature: CallController#dial now defaults the outbound caller ID to that of the controller's call
+  * Change: The command to take control of a call is now 'take' rather than 'use'.  'take' called without a call ID present a list of currently running calls
 
 # 2.0.0.alpha2 - 2012-01-30
   * Change: Plugins no longer load dialplan/event/rpc/console methods using corresponding class methods